debug
I tried not using the ESP32's IO to generate test signals directly, but instead used a variable resistor. 
The results seem to work normally, but if the ADC pin is left disconnected, it still shows a sine wave.
Suspected it might be a voltage issue.
It seems to be a pin issue. After switching to output through IO2, it worked normally.